Sync with master

+++ home.admin/97addMobileWalletLNDconnect.sh
@@ -27,17 +27,49 @@ if [ ${#GOPATH} -eq 0 ]; then
   export PATH=$PATH:$GOPATH/bin
 fi
 
+# get cpu architecture
+isARM=$(uname -m | grep -c 'arm')
+isAARCH64=$(uname -m | grep -c 'aarch64')
+isX86_64=$(uname -m | grep -c 'x86_64')
+isX86_32=$(uname -m | grep -c 'i386\|i486\|i586\|i686\|i786')
+
 # make sure go is installed
-goVersion="1.11"
+goVersion="1.12.8"
 echo "### Check Framework: GO ###"
 goInstalled=$(go version 2>/dev/null | grep -c 'go')
 if [ ${goInstalled} -eq 0 ];then
-  echo "---> Installing GO"
-  wget https://storage.googleapis.com/golang/go${goVersion}.linux-armv6l.tar.gz
-  sudo tar -C /usr/local -xzf go${goVersion}.linux-armv6l.tar.gz
+  goVersion="1.12.8"
+  if [ ${isARM} -eq 1 ] ; then
+    goOSversion="armv6l"
+  fi
+  if [ ${isAARCH64} -eq 1 ] ; then
+    goOSversion="arm64"
+  fi
+  if [ ${isX86_64} -eq 1 ] ; then
+    goOSversion="amd64"
+  fi 
+  if [ ${isX86_32} -eq 1 ] ; then
+    goOSversion="386"
+  fi 
+
+  echo "*** Installing Go v${goVersion} for ${goOSversion} ***"
+
+  # wget https://storage.googleapis.com/golang/go${goVersion}.linux-${goOSversion}.tar.gz
+  wget https://dl.google.com/go/go${goVersion}.linux-${goOSversion}.tar.gz
+  if [ ! -f "./go${goVersion}.linux-${goOSversion}.tar.gz" ]
+  then
+      echo "!!! FAIL !!! Download not success."
+      exit 1
+  fi
+  sudo tar -C /usr/local -xzf go${goVersion}.linux-${goOSversion}.tar.gz
   sudo rm *.gz
   sudo mkdir /usr/local/gocode
   sudo chmod 777 /usr/local/gocode
+  export GOROOT=/usr/local/go
+  export PATH=$PATH:$GOROOT/bin
+  export GOPATH=/usr/local/gocode
+  export PATH=$PATH:$GOPATH/bin
+  sudo bash -c "echo 'PATH=\$PATH:/usr/local/gocode/bin/' >> /etc/profile"
   goInstalled=$(go version 2>/dev/null | grep -c 'go')
 fi
 if [ ${goInstalled} -eq 0 ];then
@@ -59,11 +91,11 @@ echo "*** Setup ***"
 echo ""
 echo "Installing lndconnect. Please wait..."
 echo ""
-echo "Getting github.com/LN-Zap/lndconnect (please wait - can take several minutes) ..."
-go get -d github.com/LN-Zap/lndconnect
-cd $GOPATH/src/github.com/LN-Zap/lndconnect
+echo "Getting github.com/rootzoll/lndconnect (please wait - can take several minutes) ..."
+go get -d github.com/rootzoll/lndconnect
+cd $GOPATH/src/github.com/rootzoll/lndconnect
 echo ""
-echo "Building github.com/LN-Zap/lndconnect ..."
+echo "Building github.com/rootzoll/lndconnect ..."
 make
 cd
 sleep 3
@@ -114,7 +146,7 @@ whiptail --backtitle "Connecting Mobile Wallet" \
 	 --title "Pairing by QR code" \
 	 --yes-button "continue" \
 	 --no-button "show QR code" \
-	 --yesno "${msg}" 20 65
+	 --yesno "${msg}" 18 65
 if [ $? -eq 1 ]; then
   lndconnect --host=${host} --port=${port}
   echo "(To shrink QR code: OSX->CMD- / LINUX-> CTRL-) Press ENTER when finished."
